backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.0
sqlalchemy[asyncio]==2.0.25
asyncpg==0.29.0
alembic==1.13.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
httpx==0.26.0
beautifulsoup4==4.12.3
openai==1.12.0
google-generativeai==0.4.0
groq==0.4.1
python-dotenv==1.0.0
pydantic==2.6.0
pydantic-settings==2.1.0
Pillow==10.2.0
